Add Container wait return type at Docker model layer
Also, improve type for the API layer.
Previously, the type accounted for a 404 response.
However, in the 404 case, a `docker.errors.NotFound` exception is raised.
