+++ stubs/docker/docker/api/container.pyi
@@ -1,8 +1,9 @@
 import datetime
 from _typeshed import Incomplete
 from typing import Literal, TypedDict, overload, type_check_only
-from typing_extensions import NotRequired, TypeAlias
+from typing_extensions import TypeAlias
 
+from docker._types import WaitContainerResponse
 from docker.types.daemon import CancellableStream
 
 from ..types import ContainerConfig, EndpointConfig, HostConfig, NetworkingConfig
@@ -14,21 +15,6 @@ class _HasId(TypedDict):
 @type_check_only
 class _HasID(TypedDict):
     ID: str
-
-@type_check_only
-class _WaitErrorDetails(TypedDict):
-    Message: str
-
-@type_check_only
-class _WaitContainerExistsResponse(TypedDict):
-    StatusCode: int
-    Error: NotRequired[_WaitErrorDetails]
-
-@type_check_only
-class _WaitNoSuchContainerErrorResponse(TypedDict):
-    message: str
-
-_WaitContainerResponseType: TypeAlias = _WaitContainerExistsResponse | _WaitNoSuchContainerErrorResponse
 
 _Container: TypeAlias = _HasId | _HasID | str
 
@@ -179,4 +165,4 @@ class ContainerApiMixin:
         container: _Container,
         timeout: int | None = None,
         condition: Literal["not-running", "next-exit", "removed"] | None = None,
-    ) -> _WaitContainerResponseType: ...
+    ) -> WaitContainerResponse: ...
